script/requirements.txt
requests==2.32.3
python-dotenv==1.1.0
sshtunnel==0.4.0
pymysql==1.1.2
paramiko==3.4.0